+++ app/Http/Controllers/ProductController.php
@@ -9,6 +9,7 @@ use App\Models\Category;
 use App\Models\restaurant;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
@@ -31,6 +32,7 @@ class ProductController extends Controller
      }
     public function index()
     {
+         Gate::authorize('Product.index');
         $product=product::where('restaurant_id',Auth::user()->restaurant_id)->get();
 
         $page='index';
@@ -42,6 +44,7 @@ class ProductController extends Controller
      */
     public function create()
     {
+          Gate::authorize('Product.create');
         $page='Create';
         $restaurant=[Auth::user()->restaurant];
         $category=Category::get();
@@ -54,6 +57,7 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
+          Gate::authorize('Product.create');
         $validate_array= $this->base->validation($request,$action='store',$id='');
         $validate=$request->validate($validate_array);
          $this->base->create_data($request->file('image'),$validate);
@@ -75,6 +79,7 @@ class ProductController extends Controller
      */
     public function edit(string $id)
     {
+        Gate::authorize('Product.Update');
         $page='Edit';
          $restaurant=[Auth::user()->restaurant];
         $category=Category::get();
@@ -87,6 +92,7 @@ class ProductController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        Gate::authorize('Product.Update');
         $validate_array= $this->base->validation($request,$action='update',$id);
         $validate=$request->validate($validate_array);
         $this->base->update_data($request->file('image'),$validate,$id);
@@ -100,6 +106,7 @@ class ProductController extends Controller
      */
     public function destroy(string $id)
     {
+        Gate::authorize('Product.delete');
 
         product::where('restaurant_id',Auth::user()->restaurant_id)->findorfail($id)->delete();
         return redirect()->route('admin.product.index')
@@ -117,6 +124,7 @@ class ProductController extends Controller
 
     public function restore(string $id)
     {
+         Gate::authorize('Product.restore');
         product::where('restaurant_id',Auth::user()->restaurant_id)->withTrashed()->findorfail($id)->restore();
         return redirect()->route('admin.product.trash')->with('msg','تمت إستعادة المطعم من سلة المحذوفات بنجاح')->with('type','info');
     }
@@ -124,6 +132,7 @@ class ProductController extends Controller
 
    public function forceDelete(string $id)
     {
+        Gate::authorize('Product.forceDelete');
         product::where('restaurant_id',Auth::user()->restaurant_id)->withTrashed()->findorfail($id)->forceDelete();
         return redirect()->route('admin.product.trash')->with('msg','تم حذف المطعم بشكل نهائي')->with('type','danger');
     }
