+++ app/Http/Controllers/RestaurantController.php
@@ -6,6 +6,7 @@ use App\Models\base;
 use App\Models\User;
 use App\Models\restaurant;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class RestaurantController extends Controller
 {
@@ -28,6 +29,7 @@ class RestaurantController extends Controller
      }
     public function index()
     {
+         Gate::authorize('Restaurant.index');
         $restaurant=restaurant::get();
 
         $page='index';
@@ -39,8 +41,9 @@ class RestaurantController extends Controller
      */
     public function create()
     {
+          Gate::authorize('Restaurant.create');
         $page='Create';
-        $users=User::get();
+        $users=User::where('type',2)->where('restaurant_id',null)->get();
         $restaurant=new restaurant();
         return view('restaurant.form',compact('page','restaurant','users'));
     }
@@ -50,6 +53,7 @@ class RestaurantController extends Controller
      */
     public function store(Request $request)
     {
+          Gate::authorize('Restaurant.create');
         $validate_array= $this->base->validation($request,$action='store',$id='');
         $validate=$request->validate($validate_array);
          $this->base->create_data($request->file('image'),$validate);
@@ -71,6 +75,7 @@ class RestaurantController extends Controller
      */
     public function edit(string $id)
     {
+        Gate::authorize('Restaurant.Update');
         $page='Edit';
         $users=User::get();
 
@@ -83,6 +88,7 @@ class RestaurantController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        Gate::authorize('Restaurant.Update');
         $validate_array= $this->base->validation($request,$action='update',$id);
         $validate=$request->validate($validate_array);
         $this->base->update_data($request->file('image'),$validate,$id);
@@ -96,8 +102,10 @@ class RestaurantController extends Controller
      */
     public function destroy(string $id)
     {
+        Gate::authorize('Restaurant.delete');
 
         restaurant::findorfail($id)->delete();
+
         return redirect()->route('admin.Restaurant.index')
          ->with('msg','تم نقل المطعم إلى سلة المحذوفات')->with('type','danger')
         ;
@@ -113,6 +121,7 @@ class RestaurantController extends Controller
 
     public function restore(string $id)
     {
+         Gate::authorize('Restaurant.restore');
         restaurant::withTrashed()->findorfail($id)->restore();
         return redirect()->route('admin.Restaurant.trash')->with('msg','تمت إستعادة المطعم من سلة المحذوفات بنجاح')->with('type','info');
     }
@@ -120,6 +129,7 @@ class RestaurantController extends Controller
 
    public function forceDelete(string $id)
     {
+        Gate::authorize('Restaurant.forceDelete');
         restaurant::withTrashed()->findorfail($id)->forceDelete();
         return redirect()->route('admin.Restaurant.trash')->with('msg','تم حذف المطعم بشكل نهائي')->with('type','danger');
     }
