+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 use App\Models\base;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -28,6 +29,7 @@ class UserController extends Controller
      }
     public function index()
     {
+         Gate::authorize('User.index');
         $user=User::get();
         $page='index';
         return view('user.index',compact('user','page'));
@@ -38,6 +40,7 @@ class UserController extends Controller
      */
     public function create()
     {
+          Gate::authorize('User.create');
         $page='Create';
         $user=new User();
         return view('user.form',compact('page','user'));
@@ -48,6 +51,7 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+          Gate::authorize('User.create');
         $validate_array= $this->base->validation($request,$action='store',$id='');
         $validate=$request->validate($validate_array);
          $this->base->create_data($image='',$validate);
@@ -70,6 +74,7 @@ class UserController extends Controller
      */
     public function edit(string $id)
     {
+        Gate::authorize('User.Update');
         $page='Edit';
         $user=User::find($id);
         return view('user.form',compact('page','user'));
@@ -80,9 +85,12 @@ class UserController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        Gate::authorize('User.Update');
         $validate_array= $this->base->validation($request,$action='update',$id);
+
         $validate=$request->validate($validate_array);
-        if(in_array('email_verified_at',$validate)){
+
+        if(in_array('email_verified_at',array_keys($validate))){
             $validate['email_verified_at']=date('Y-m-d H:i:s');
         }else{
             $validate['email_verified_at']=null;
@@ -97,7 +105,13 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
-        User::findorfail($id)->delete();
+        Gate::authorize('User.delete');
+        $user=User::findorfail($id);
+        $user->restaurant->update([
+         'user_id'=>null
+        ]);
+        $user->delete();
+
         return redirect()->route('admin.user.index')
          ->with('msg','تم نقل المطعم إلى سلة المحذوفات')->with('type','danger')
         ;
@@ -113,6 +127,7 @@ class UserController extends Controller
 
     public function restore(string $id)
     {
+         Gate::authorize('User.restore');
         User::withTrashed()->findorfail($id)->restore();
         return redirect()->route('admin.user.trash')->with('msg','تمت إستعادة المطعم من سلة المحذوفات بنجاح')->with('type','info');
     }
@@ -120,6 +135,7 @@ class UserController extends Controller
 
    public function forceDelete(string $id)
     {
+        Gate::authorize('User.forceDelete');
         User::withTrashed()->findorfail($id)->forceDelete();
         return redirect()->route('admin.user.trash')->with('msg','تم حذف المطعم بشكل نهائي')->with('type','danger');
     }
